rename script, combine into one
